Migrate and simplify more tests.
