+++ cqf-fhir-benchmark/src/test/java/org/opencds/cqf/fhir/benchmark/MeasuresAdditionalData.java
@@ -47,7 +47,7 @@ public class MeasuresAdditionalData {
 
         Bundle additionalData = (Bundle) FhirContext.forR4Cached()
                 .newJsonParser()
-                .parseResource(this.getClass().getResourceAsStream("CaseRepresentation101/generated.json"));
+                .parseResource(this.getClass().getResourceAsStream("CaseRepresentation101/additionalData.json"));
 
         this.when = Measure.given()
                 .repositoryFor("CaseRepresentation101")
