+++ cqf-fhir-cr/src/test/java/org/opencds/cqf/fhir/cr/measure/r4/MeasureProcessorEvaluateTest.java
@@ -4,170 +4,15 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.rest.server.exceptions.InvalidRequestException;
-import java.text.SimpleDateFormat;
-import java.time.LocalDate;
-import java.time.Month;
-import java.time.ZoneId;
-import java.time.ZoneOffset;
-import java.util.HashMap;
-import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.MeasureReport;
-import org.hl7.fhir.r4.model.Period;
 import org.junit.jupiter.api.Test;
-import org.opencds.cqf.fhir.cql.engine.retrieve.RetrieveSettings.SEARCH_FILTER_MODE;
-import org.opencds.cqf.fhir.cql.engine.retrieve.RetrieveSettings.TERMINOLOGY_FILTER_MODE;
-import org.opencds.cqf.fhir.cql.engine.terminology.TerminologySettings.VALUESET_EXPANSION_MODE;
-import org.opencds.cqf.fhir.cr.measure.MeasureEvaluationOptions;
 import org.opencds.cqf.fhir.cr.measure.r4.Measure.Given;
-import org.opencds.cqf.fhir.cr.measure.r4.utils.R4DateHelper;
 
 class MeasureProcessorEvaluateTest {
 
-    private static final SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd");
-
     protected static Given given = Measure.given().repositoryFor("CaseRepresentation101");
 
-    @Test
-    void measure_eval() {
-
-        var start = LocalDate.of(2022, Month.JANUARY, 1).atStartOfDay(ZoneId.systemDefault());
-        var end = LocalDate.of(2022, Month.JUNE, 29).atStartOfDay(ZoneId.systemDefault());
-        var helper = new R4DateHelper();
-        var measurementPeriod = helper.buildMeasurementPeriod(start, end);
-        var report = given.when()
-                .measureId("GlycemicControlHypoglycemicInitialPopulation")
-                .periodStart(start)
-                .periodEnd(end)
-                .subject("Patient/eNeMVHWfNoTsMTbrwWQQ30A3")
-                .reportType("subject")
-                .evaluate()
-                .then()
-                .report();
-
-        assertEquals(
-                measurementPeriod.getStart().toInstant(),
-                report.getPeriod().getStart().toInstant());
-        assertEquals(
-                measurementPeriod.getEnd().toInstant(),
-                report.getPeriod().getEnd().toInstant());
-    }
-
-    @Test
-    void measure_eval_UTC() {
-        var start = LocalDate.of(2022, Month.JANUARY, 1).atStartOfDay(ZoneOffset.UTC);
-        var end = LocalDate.of(2022, Month.JUNE, 29).atStartOfDay(ZoneOffset.UTC);
-        var helper = new R4DateHelper();
-        var measurementPeriod = helper.buildMeasurementPeriod(start, end);
-        var report = given.when()
-                .measureId("GlycemicControlHypoglycemicInitialPopulation")
-                .periodStart(start)
-                .periodEnd(end)
-                .subject("Patient/eNeMVHWfNoTsMTbrwWQQ30A3")
-                .reportType("subject")
-                .evaluate()
-                .then()
-                .report();
-
-        final Period actualPeriod = report.getPeriod();
-
-        assertEquals(
-                measurementPeriod.getStart().toInstant(),
-                actualPeriod.getStart().toInstant());
-        assertEquals(
-                measurementPeriod.getEnd().toInstant(), actualPeriod.getEnd().toInstant());
-    }
-
-    @Test
-    void measure_eval_with_additional_data() {
-        Bundle additionalData = (Bundle) FhirContext.forR4Cached()
-                .newJsonParser()
-                .parseResource(
-                        MeasureProcessorEvaluateTest.class.getResourceAsStream("CaseRepresentation101/generated.json"));
-        var report = given.when()
-                .measureId("GlycemicControlHypoglycemicInitialPopulation")
-                .periodStart("2022-01-01")
-                .periodEnd("2022-01-31")
-                .subject("Patient/980babd9-4979-4b76-978c-946719022dbb")
-                .additionalData(additionalData)
-                .evaluate()
-                .then()
-                .hasMeasureVersion("0.000.01")
-                .report();
-
-        assertEquals("2022-01-01", formatter.format(report.getPeriod().getStart()));
-        assertEquals("2022-01-31", formatter.format(report.getPeriod().getEnd()));
-    }
-
-    @Test
-    void with_custom_options() {
-        var evaluationOptions = MeasureEvaluationOptions.defaultOptions();
-        evaluationOptions.getEvaluationSettings().setLibraryCache(new HashMap<>());
-        evaluationOptions
-                .getEvaluationSettings()
-                .getRetrieveSettings()
-                .setSearchParameterMode(SEARCH_FILTER_MODE.FILTER_IN_MEMORY)
-                .setTerminologyParameterMode(TERMINOLOGY_FILTER_MODE.FILTER_IN_MEMORY);
-
-        evaluationOptions
-                .getEvaluationSettings()
-                .getTerminologySettings()
-                .setValuesetExpansionMode(VALUESET_EXPANSION_MODE.PERFORM_NAIVE_EXPANSION);
-
-        var when = Measure.given()
-                .repositoryFor("CaseRepresentation101")
-                .evaluationOptions(evaluationOptions)
-                .when()
-                .measureId("GlycemicControlHypoglycemicInitialPopulation")
-                .subject("Patient/eNeMVHWfNoTsMTbrwWQQ30A3")
-                .periodStart("2022-01-01")
-                .periodEnd("2022-06-29")
-                .reportType("subject")
-                .evaluate();
-
-        assertNotNull(when.then().report());
-
-        // Run again to find any issues with caching
-        when.then().report();
-    }
-
-    @Test
-    void additional_data_with_custom_options() {
-        var evaluationOptions = MeasureEvaluationOptions.defaultOptions();
-        evaluationOptions.getEvaluationSettings().setLibraryCache(new HashMap<>());
-        evaluationOptions
-                .getEvaluationSettings()
-                .getRetrieveSettings()
-                .setSearchParameterMode(SEARCH_FILTER_MODE.FILTER_IN_MEMORY)
-                .setTerminologyParameterMode(TERMINOLOGY_FILTER_MODE.FILTER_IN_MEMORY);
-
-        evaluationOptions
-                .getEvaluationSettings()
-                .getTerminologySettings()
-                .setValuesetExpansionMode(VALUESET_EXPANSION_MODE.PERFORM_NAIVE_EXPANSION);
-
-        Bundle additionalData = (Bundle) FhirContext.forR4Cached()
-                .newJsonParser()
-                .parseResource(
-                        MeasureProcessorEvaluateTest.class.getResourceAsStream("CaseRepresentation101/generated.json"));
-
-        var when = Measure.given()
-                .repositoryFor("CaseRepresentation101")
-                .evaluationOptions(evaluationOptions)
-                .when()
-                .measureId("GlycemicControlHypoglycemicInitialPopulation")
-                .periodStart("2022-01-01")
-                .periodEnd("2022-01-31")
-                .subject("Patient/980babd9-4979-4b76-978c-946719022dbb")
-                .additionalData(additionalData)
-                .evaluate();
-
-        assertNotNull(when.then().report());
-
-        // Run again to find any issues with caching
-        when.then().report();
-    }
     /**
      * test to validate that measure with MeasureScorer specified at the group level
      * and nothing on measure-level MeasureScorer
