+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/AdditionalDataAndCustomOptions/resources/library/GlycemicControlHypoglycemicInitialPopulation.json
@@ -1,0 +1,28 @@
+{
+  "resourceType": "Library",
+  "id": "GlycemicControlHypoglycemicInitialPopulation",
+  "meta": {
+    "profile": [ "http://hl7.org/fhir/us/cqfmeasures/StructureDefinition/library-cqfm" ]
+  },
+  "extension": [ {
+    "url": "http://hl7.org/fhir/us/cqfmeasures/StructureDefinition/cqfm-populationBasis",
+    "valueCode": "Encounter"
+  } ],
+  "url": "http://test-case.com/fhir/measures/Library/GlycemicControlHypoglycemicInitialPopulation",
+  "version": "0.0.004",
+  "name": "GlycemicControlHypoglycemicInitialPopulation",
+  "status": "draft",
+  "experimental": false,
+  "type": {
+    "coding": [ {
+      "system": "http://terminology.hl7.org/CodeSystem/library-type",
+      "code": "logic-library",
+      "display": "Logic Library"
+    } ]
+  },
+  "description": "All inpatient encounters (including ED/Observation visits that end within 1 hour of the start of the inpatient encounter) for patients of all ages where at least one ADD was ordered or administered during the encounter that is during the measurement period.",
+  "content": [ {
+    "contentType": "text/cql",
+    "url": "../../cql/GlycemicControlHypoglycemicInitialPopulation.cql"
+  } ]
+}
