+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/AdditionalDataAndCustomOptions/valueset/2.16.840.1.113762.1.4.1.json
@@ -1,0 +1,40 @@
+{
+  "resourceType": "ValueSet",
+  "id": "2.16.840.1.113762.1.4.1",
+  "meta": {
+    "versionId": "5",
+    "lastUpdated": "2015-03-31T01:00:01.000-04:00"
+  },
+  "url": "http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113762.1.4.1",
+  "version": "20150331",
+  "name": "ONC Administrative Sex",
+  "title": "ONC Administrative Sex",
+  "status": "active",
+  "date": "2015-03-31T01:00:01-04:00",
+  "publisher": "ONC",
+  "description": "Gender identity restricted to only Male and Female used in administrative situations requiring a restriction to these two categories. ",
+  "expansion": {
+    "identifier": "urn:uuid:7c494f5c-d0b0-48b6-9612-99bc4ce6969c",
+    "timestamp": "2022-03-09T09:21:02-05:00",
+    "total": 2,
+    "offset": 0,
+    "parameter": [ {
+      "name": "count",
+      "valueInteger": 1000
+    }, {
+      "name": "offset",
+      "valueInteger": 0
+    } ],
+    "contains": [ {
+      "system": "http://terminology.hl7.org/CodeSystem/v3-AdministrativeGender",
+      "version": "20210301",
+      "code": "F",
+      "display": "Female"
+    }, {
+      "system": "http://terminology.hl7.org/CodeSystem/v3-AdministrativeGender",
+      "version": "20210301",
+      "code": "M",
+      "display": "Male"
+    } ]
+  }
+}
