+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/AdditionalDataAndCustomOptions/valueset/encounter-discharge-disposition.json
@@ -1,0 +1,147 @@
+{
+  "resourceType": "ValueSet",
+  "id": "encounter-discharge-disposition",
+  "meta": {
+    "lastUpdated": "2020-02-23T18:41:39.109-07:00",
+    "profile": [
+      "http://hl7.org/fhir/StructureDefinition/shareablevalueset"
+    ]
+  },
+  "url": "http://terminology.hl7.org/ValueSet/encounter-discharge-disposition",
+  "identifier": [
+    {
+      "system": "urn:ietf:rfc:3986",
+      "value": "urn:oid:2.16.840.1.113883.4.642.3.259"
+    }
+  ],
+  "version": "0.3.0",
+  "name": "DischargeDisposition",
+  "title": "Discharge disposition",
+  "status": "draft",
+  "experimental": false,
+  "date": "2020-02-24T12:41:39+11:00",
+  "publisher": "FHIR Project team",
+  "contact": [
+    {
+      "telecom": [
+        {
+          "system": "url",
+          "value": "http://hl7.org/fhir"
+        }
+      ]
+    }
+  ],
+  "description": "This value set defines a set of codes that can be used to where the patient left the hospital.",
+  "compose": {
+    "include": [
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "concept": [
+          {
+            "code": "home",
+            "display": "Home"
+          },
+          {
+            "code": "alt-home",
+            "display": "Alternative home"
+          },
+          {
+            "code": "other-hcf",
+            "display": "Other healthcare facility"
+          },
+          {
+            "code": "hosp",
+            "display": "Hospice"
+          },
+          {
+            "code": "long",
+            "display": "Long-term care"
+          },
+          {
+            "code": "aadvice",
+            "display": "Left against advice"
+          },
+          {
+            "code": "exp",
+            "display": "Expired"
+          },
+          {
+            "code": "psy",
+            "display": "Psychiatric hospital"
+          },
+          {
+            "code": "rehab",
+            "display": "Rehabilitation"
+          },
+          {
+            "code": "snf",
+            "display": "Skilled nursing facility"
+          },
+          {
+            "code": "oth",
+            "display": "Other"
+          }
+        ]
+      }
+    ]
+  },
+  "expansion": {
+    "contains": [
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "home",
+        "display": "Home"
+      },
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "alt-home",
+        "display": "Alternative home"
+      },
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "other-hcf",
+        "display": "Other healthcare facility"
+      },
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "hosp",
+        "display": "Hospice"
+      },
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "long",
+        "display": "Long-term care"
+      },
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "aadvice",
+        "display": "Left against advice"
+      },
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "exp",
+        "display": "Expired"
+      },
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "psy",
+        "display": "Psychiatric hospital"
+      },
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "rehab",
+        "display": "Rehabilitation"
+      },
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "snf",
+        "display": "Skilled nursing facility"
+      },
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/discharge-disposition",
+        "code": "oth",
+        "display": "Other"
+      }
+    ]
+  }
+}
